+++ priv/repo/migrations/20191218211301_create_transactions.exs
@@ -3,9 +3,9 @@ defmodule Bank.Repo.Migrations.CreateTransactions do
 
   def change do
     create table(:transactions) do
-      add :amount, :decimal, null: false
-      add :prev_account_balance, :decimal, null: false
-      add :new_account_balance, :decimal, null: false
+      add :amount, :float, null: false
+      add :prev_account_balance, :float, null: false
+      add :new_account_balance, :float, null: false
       add :operation_id, references(:operations), null: false
       add :account_id, references(:accounts), null: false
       add :target_account_id, references(:accounts)
